MotionAppFiles/requirements.txt
altgraph==0.17.4
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
et_xmlfile==2.0.0
graphviz==0.20.3
idna==3.10
macholib==1.16.3
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
setuptools==75.8.2
six==1.17.0
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wheel==0.45.1